+++ bin/clear_data.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 '''clear_data.py
 
 Description: