+++ bin/convertPIMCuuid.py
@@ -1,4 +1,5 @@
-#!/usr/bin/python3
+#!/usr/bin/env python
+
 # Filename: convertPIMCuuid.py
 # Author: Nathan Nichols
 #